Added authentication to roasterController

+++ controllers/roasterController.js
@@ -5,6 +5,24 @@ const jwt = require("jsonwebtoken");
 const mongoose = require("mongoose");
 const Roaster = require("../models/roasterModel")
 
+const checkAuthStatus = request => {
+    if (!request.headers.authorization) {
+        return false
+    }
+    const token = request.headers.authorization.split(" ")[1]
+
+    const loggedInUser = jwt.verify(token, process.env.JWT_SECRET, (err, data) => {
+        if (err) {
+            return false
+        }
+        else {
+            return data
+        }
+    });
+    console.log(loggedInUser)
+    return loggedInUser
+}
+
 // Get all roasters
 router.get("/api/roasters", async function (req, res) {
     try {
@@ -48,6 +66,11 @@ router.get("/api/roasters/search/:name", async function (req, res) {
 
 // Add a roaster
 router.post("/api/roasters", async function (req, res) {
+    const loggedInUser = checkAuthStatus(req);
+    if(!loggedInUser){
+        return res.status(401).send("Must be logged in")
+    }
+    console.log(loggedInUser);
     try {
         let newRoaster = await Roaster.create({
             name: req.body.name,
@@ -64,6 +87,11 @@ router.post("/api/roasters", async function (req, res) {
 
 // Edit a roaster
 router.put("/api/roasters/:id", async function (req, res) {
+    const loggedInUser = checkAuthStatus(req);
+    if(!loggedInUser){
+        return res.status(401).send("Must be logged in")
+    }
+    console.log(loggedInUser);
     try {
         if (req.body.cafeId) {
             await Roaster.findOneAndUpdate({
@@ -95,6 +123,11 @@ router.put("/api/roasters/:id", async function (req, res) {
 
 // Delete a roaster
 router.delete("/api/roasters/:id", async function (req, res) {
+    const loggedInUser = checkAuthStatus(req);
+    if(!loggedInUser){
+        return res.status(401).send("Must be logged in")
+    }
+    console.log(loggedInUser);
     try {
         let result = await Roaster.deleteOne({ _id: mongoose.Types.ObjectId(req.params.id) })
         res.json(result)
